src/finley/requirements/requirements.txt
azure-kusto-data==5.0.2
azure-identity
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-search-documents==11.6.0b7
azure-ai-contentsafety==1.0.0
azure-ai-evaluation==1.5.0
azure-ai-inference==1.0.0b6
azure-ai-projects==1.0.0b8
bs4==0.0.1
fastapi==0.109.2
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-langchain==0.39.0
opentelemetry-instrumentation-openai==0.39.0
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
pandas>=2.0.0
pydantic
pyodbc==5.2.0
python-dotenv
tavily-python==0.7.0
uvicorn[standard]==0.27.1
